reduces code
